Add support for service host factory methods.
-Add ServiceHostFactoryMethodWrapper to wrap the factory method.
-Remove dependency on custom IServiceHostFactory implementation. Now can simply supply a lamda expression represending a method that can create a ServiceHost object.
