+++ Hyper.WcfHosting/Hyper.WcfHosting.csproj
@@ -46,6 +46,7 @@
     <Compile Include="IServiceHostFactory.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="HyperServiceHostContainer.cs" />
+    <Compile Include="ServiceHostFactoryMethodWrapper.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
