+++ HyperNetExtensibilityTest/HyperNetExtensibilityTest.csproj
@@ -43,7 +43,6 @@
     <Compile Include="ActivityMonitors\DatabaseActivityMonitor.cs" />
     <Compile Include="CommandModules\DisposableCommandModule.cs" />
     <Compile Include="CommandModules\EchoCommandModule.cs" />
-    <Compile Include="WcfHosting\HyperNodeServiceHostFactory.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ActivityMonitors\TaskActivityTracer.cs" />
   </ItemGroup>
